Don't store event_id in the session
Instead, make it a page parameter.  This does complicate things a
bit, since most pages require the event id, but it allows you to
have multiple tabs open for separate events, which the previous
code did not allow.

While here, stop echoing gift, senders, and receivers in flash
message on successful gift add.  These are now shown in the Recent
Gifts table, so there is no reason to clutter the flash.